FreeLancerCTF/web/captcha1/content/requirements.txt
click==6.6
Flask==0.11.1
Flask-WTF==0.13.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.4.1
Werkzeug==0.11.11
WTForms==2.1
